Develop (#79)
* Bug correction in sienaDataCreate

Issue #75 is hopefully solved by this change.

* Effects object - version and comparisons

Another modification in view of Issue #75. The comparison between effects objects in initializeFRAN is now based on the shortName instead of the effectName. Additionally, the effects object is given a "version" attribute.

* Version 1.3.23

Various changes, see NEWS.md.

* Edit manual

A short addition was made to the manual. Furthermore, the superfluous C++ class `DifferentCovariatePredicate` was dropped.

* Corrections of version 1.3.23

Corrections of attributes of effects object and input of list of sienaFit objects for sienaGOF.

* Revert "Corrections of version 1.3.23"

This reverts commit bddedf3f3012b6965d590829447701ac31180fbd.